Rajout de certains checks pour le deploiement de wapt
Subversion-branch: /trunk
Subversion-revision: 758

+++ waptserver/waptserver.py
@@ -43,6 +43,7 @@ import ConfigParser
 import logging
 import codecs
 import zipfile
+import platform
 import pprint
 import socket
 import requests
@@ -182,11 +183,11 @@ def informations():
     if os.path.exists(waptsetup):
         pe = pefile.PE(waptsetup)
         informations["client_version"] =  pe.FileInfo[0].StringTable[0].entries['ProductVersion'].strip()  
-    
+
     return  Response(response=json.dumps(informations),
-                    status=200,
-                    mimetype="application/json")    
-    
+                     status=200,
+                     mimetype="application/json")    
+
 
 @app.route('/wapt/')
 def wapt_listing():
@@ -232,8 +233,8 @@ def get_host_list():
             list_hosts.append(host)
 
     return  Response(response=json.dumps(list_hosts),
-                    status=200,
-                    mimetype="application/json")
+                     status=200,
+                     mimetype="application/json")
 
 @app.route('/update_host',methods=['POST'])
 def update_host():
@@ -277,8 +278,8 @@ def get_client_software_list(uuid=""):
     softwares = get_host_data(uuid, filter={"softwares":1})
     if softwares.has_key('softwares'):
         return  Response(response=json.dumps(softwares['softwares']),
-                    status=200,
-                    mimetype="application/json")
+                         status=200,
+                         mimetype="application/json")
     else:
         return "{}"
 
@@ -323,12 +324,12 @@ def get_client_package_list(uuid=""):
             if matching:
                 if package < matching[-1]:
                     p['install_status'] = 'NEED-UPGRADE'
-    
+
         return  Response(response=json.dumps(packages['packages']),
-                        status=200,
-                        mimetype="application/json")
+                         status=200,
+                         mimetype="application/json")
     return "{}"
-    
+
 
 def requires_auth(f):
     @wraps(f)
@@ -440,18 +441,18 @@ def get_hosts_by_group(name=""):
         package = PackageEntry()     
         for h in hosts:  
             package.load_control_from_wapt(h)
-            
+
             if name in package.depends.split(','):            
                 list_hosts.append({"computer_fqdn":package.package})            
-            
+
         return  Response(response=json.dumps(list_hosts),
-                                status=200,
-                                mimetype="application/json")                               
+                         status=200,
+                         mimetype="application/json")                               
     except:
         e = sys.exc_info()
         return str(e)    
     return "Unsupported method"    
-    
+
 @app.route('/upgrade_host/<string:ip>')
 def upgrade_host(ip):
     try:
@@ -474,15 +475,15 @@ def upgrade_host(ip):
 def install_wapt(computer_name,authentication_file):
     cmd = '/usr/bin/smbclient -G -E -A %s  //%s/IPC$ -c listconnect ' % (authentication_file, computer_name)
     try:
-    	subprocess.check_output(cmd,stderr=subprocess.STDOUT,shell=True)
+        subprocess.check_output(cmd,stderr=subprocess.STDOUT,shell=True)
     except subprocess.CalledProcessError as e:
         if "NT_STATUS_LOGON_FAILURE" in e.output:
-	    raise Exception("Mauvais identifiants")
-	if "NT_STATUS_CONNECTION_REFUSED" in e.output:	
-	    raise Exception("Partage IPC$ non accessible")
+            raise Exception("Mauvais identifiants")
+        if "NT_STATUS_CONNECTION_REFUSED" in e.output:	
+            raise Exception("Partage IPC$ non accessible")
 
-	raise Exception(u"%s" % e.output)
-	
+        raise Exception(u"%s" % e.output)
+
 
     cmd = '/usr/bin/smbclient -A "%s" //%s/c\\$ -c "put waptsetup.exe" ' % (authentication_file, computer_name)
     print subprocess.check_output(cmd,shell=True)
@@ -500,41 +501,48 @@ def install_wapt(computer_name,authentication_file):
 
     cmd = '/usr/bin/winexe -A "%s"  //%s  "c:\\wapt\\wapt-get.exe --version"' % (authentication_file, computer_name)
     return subprocess.check_output(cmd,shell=True)
-        
+
 @app.route('/deploy_wapt',methods=['POST'])
 def deploy_wapt():
     try:
-	result = {}
-	if request.method == 'POST':
-	    d = json.loads(request.data) 
-	    if not d.has_key('auth'):  
-		raise Exception("Les informations d'authentification sont manquantes")
-	    if not d.has_key('computer_fqdn'):  
-		raise Exception(u"Il n'y a aucuns ordinateurs de renseigné")
-	    
-	    auth_file = tempfile.mkstemp("wapt")[1]
-	    try:
-		with open(auth_file, 'w') as f:
-		    f.write('username = %s\npassword = %s\ndomain = %s\n'% (
-			d['auth']['username'],
-			d['auth']['password'],
-			d['auth']['domain']))
-		    
-		os.chdir(wapt_folder)
-		
-		message = install_wapt(d['computer_fqdn'],auth_file)
-					       
-		result = { 'status' : 'OK' , 'message': message} 
-	    finally:
-		os.unlink(auth_file)
-	    
-	else:
-	    raise Exception(u"methode http non supportée")
+        result = {}
+        if platform.system() != 'Linux':
+            raise Exception(u'Le serveur wapt doit être executé sous Linux')
+        if subprocess.call('which smbclient',shell=True) != 0:
+            raise Exception(u"smbclient n'est pas installé sur le serveur wapt")
+        if subprocess.call('which winexe',shell=True) != 0:
+            raise Exception(u"winexe n'est pas installé sur le serveur wapt")
+            
+        if request.method == 'POST':
+            d = json.loads(request.data) 
+            if not d.has_key('auth'):  
+                raise Exception("Les informations d'authentification sont manquantes")
+            if not d.has_key('computer_fqdn'):  
+                raise Exception(u"Il n'y a aucuns ordinateurs de renseigné")
+
+            auth_file = tempfile.mkstemp("wapt")[1]
+            try:
+                with open(auth_file, 'w') as f:
+                    f.write('username = %s\npassword = %s\ndomain = %s\n'% (
+                        d['auth']['username'],
+                        d['auth']['password'],
+                        d['auth']['domain']))
+
+                os.chdir(wapt_folder)
+
+                message = install_wapt(d['computer_fqdn'],auth_file)
+
+                result = { 'status' : 'OK' , 'message': message} 
+            finally:
+                os.unlink(auth_file)
+
+        else:
+            raise Exception(u"methode http non supportée")
 
     except Exception, e:
-	result = { 'status' : 'ERROR', 'message': u"%s" % e  }
+        result = { 'status' : 'ERROR', 'message': u"%s" % e  }
 
-	
+
     return json.dumps(result)
 
 @app.route('/login',methods=['POST'])
@@ -628,9 +636,9 @@ def check_auth(username, password):
 def authenticate():
     """Sends a 401 response that enables basic auth"""
     return Response(
-    'Could not verify your access level for that URL.\n'
-    'You have to login with proper credentials', 401,
-    {'WWW-Authenticate': 'Basic realm="Login Required"'})
+        'Could not verify your access level for that URL.\n'
+        'You have to login with proper credentials', 401,
+        {'WWW-Authenticate': 'Basic realm="Login Required"'})
 
 if __name__ == "__main__":
     debug=False
@@ -645,4 +653,3 @@ if __name__ == "__main__":
         except KeyboardInterrupt:
             logger.info("stopping waptserver")
             server.stop()
-
